requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
boto3==1.34.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.0
sqlalchemy==2.0.23
pandas==2.1.4
numpy==1.26.2
pytest==7.4.3
requests==2.31.0
